frontend: Split the routing-related stuff in multiple files under routing/
