+++ frontend/src/components/OAuth2Session.tsx
@@ -18,8 +18,8 @@ import { atomFamily } from "jotai/utils";
 import { atomWithMutation } from "jotai-urql";
 import { useTransition } from "react";
 
-import { Link } from "../Router";
 import { FragmentType, graphql, useFragment } from "../gql";
+import { Link } from "../routing";
 import { getDeviceIdFromScope } from "../utils/deviceIdFromScope";
 
 // import LoadingSpinner from "./LoadingSpinner/LoadingSpinner";
