+++ frontend/src/components/UserEmail/UserEmail.tsx
@@ -19,8 +19,8 @@ import { atomFamily } from "jotai/utils";
 import { atomWithMutation } from "jotai-urql";
 import { useTransition } from "react";
 
-import { Link } from "../../Router";
 import { FragmentType, graphql, useFragment } from "../../gql";
+import { Link } from "../../routing";
 
 import styles from "./UserEmail.module.css";
