+++ frontend/src/main.tsx
@@ -17,10 +17,12 @@ import { DevTools } from "jotai-devtools";
 import { Suspense, StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 
-import Router from "./Router";
 import { HydrateAtoms } from "./atoms";
 import LoadingScreen from "./components/LoadingScreen";
+import { Router } from "./routing";
 import "./main.css";
+import LoadingSpinner from "./components/LoadingSpinner";
+import Layout from "./components/Layout";
 
 createRoot(document.getElementById("root") as HTMLElement).render(
   <StrictMode>
@@ -28,7 +30,11 @@ createRoot(document.getElementById("root") as HTMLElement).render(
       {import.meta.env.DEV && <DevTools />}
       <HydrateAtoms>
         <Suspense fallback={<LoadingScreen />}>
-          <Router />
+          <Layout>
+            <Suspense fallback={<LoadingSpinner />}>
+              <Router />
+            </Suspense>
+          </Layout>
         </Suspense>
       </HydrateAtoms>
     </Provider>
