+++ frontend/src/routing/index.ts
@@ -1,0 +1,20 @@
+// Copyright 2023 The Matrix.org Foundation C.I.C.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+export { default as Router } from "./Router";
+export { default as Link } from "./Link";
+export type { Route, Location } from "./routes";
+export { pathToRoute, routeToPath } from "./routes";
+export { getRouteActionRedirection } from "./actions";
+export { routeAtom, locationAtom, appConfigAtom } from "./atoms";
