+++ frontend/src/routing/routes.ts
@@ -39,3 +39,114 @@ export type Route =
   | CompatSessionListRoute
   | VerifyEmailRoute
   | UnknownRoute;
+
+export type Segments = string[];
+
+// Converts a list of segments to a path
+const segmentsToPath = (segments: Segments): string =>
+  segments.map((part) => encodeURIComponent(part)).join("/");
+
+// Converts a path to a list of segments
+const pathToSegments = (path: string): Segments =>
+  path.split("/").map(decodeURIComponent);
+
+// Converts a route to a path
+export const routeToPath = (route: Route): string =>
+  segmentsToPath(routeToSegments(route));
+
+// Converts a path to a route
+export const pathToRoute = (path: string): Route =>
+  segmentsToRoute(pathToSegments(path));
+
+// Converts a route to a list of segments
+export const routeToSegments = (route: Route): Segments => {
+  switch (route.type) {
+    case "profile":
+      return [];
+    case "sessions-overview":
+      return ["sessions-overview"];
+    case "session":
+      return ["session", route.id];
+    case "verify-email":
+      return ["emails", route.id, "verify"];
+    case "client":
+      return ["clients", route.id];
+    case "browser-session-list":
+      return ["browser-sessions"];
+    case "browser-session":
+      return ["browser-sessions", route.id];
+    case "oauth2-session-list":
+      return ["oauth2-sessions"];
+    case "compat-session-list":
+      return ["compat-sessions"];
+    case "unknown":
+      return route.segments;
+  }
+};
+
+const P = Symbol();
+type PatternItem = string | typeof P;
+
+// Returns true if the segments match the pattern, where P is a parameter
+const segmentMatches = (
+  segments: Segments,
+  ...pattern: PatternItem[]
+): boolean => {
+  // Quick check to see if the lengths match
+  if (segments.length !== pattern.length) return false;
+
+  // Check each segment
+  for (let i = 0; i < segments.length; i++) {
+    // If the pattern is P, then it's a parameter and we can skip it
+    if (pattern[i] === P) continue;
+    // Otherwise, check that the segment matches the pattern
+    if (segments[i] !== pattern[i]) return false;
+  }
+
+  return true;
+};
+
+// Converts a list of segments to a route
+export const segmentsToRoute = (segments: Segments): Route => {
+  const matches = (...pattern: PatternItem[]): boolean =>
+    segmentMatches(segments, ...pattern);
+
+  // Special case for the home page
+  if (matches() || matches("")) {
+    return { type: "profile" };
+  }
+
+  if (matches("sessions-overview")) {
+    return { type: "sessions-overview" };
+  }
+
+  if (matches("browser-sessions")) {
+    return { type: "browser-session-list" };
+  }
+
+  if (matches("oauth2-sessions")) {
+    return { type: "oauth2-session-list" };
+  }
+
+  if (matches("compat-sessions")) {
+    return { type: "compat-session-list" };
+  }
+
+  if (matches("emails", P, "verify")) {
+    return { type: "verify-email", id: segments[1] };
+  }
+
+  if (matches("clients", P)) {
+    return { type: "client", id: segments[1] };
+  }
+
+  if (matches("browser-sessions", P)) {
+    return { type: "browser-session", id: segments[1] };
+  }
+
+  if (matches("session", P)) {
+    return { type: "session", id: segments[1] };
+  }
+
+  return { type: "unknown", segments };
+};
